+++ RELEASE_NOTES.md
@@ -3,12 +3,17 @@
 ## Version 0.1.12
 
 ### Breaking Changes
-- 
+- ExplainerDashboardStandaloneTab is now called ExplainerTab
 - 
 
 ### New Features
--
--
+
+added support for the `jupyter-dash` package for inline dashboard in 
+Jupyter notebooks, adding the following dashboard classes:
+
+- `JupyterExplainerDashboard`
+- `JupyterExplainerTab`
+- `InlineExplainer`
 
 ### Bug Fixes
 -
@@ -21,6 +26,7 @@
 ### Other Changes
 -
 -
+
 
 ## Template:
 ### Breaking Changes